Set up forecast fragment to display API data
It is currently pre-set to display Vilnius current weather. API data replaces mocked data in forecast fragment

+++ app/src/main/java/lt/paulius/weatherforecastapplication/repository/Current.kt
@@ -1,3 +1,3 @@
 package lt.paulius.weatherforecastapplication.repository
 
-data class Current(val temp_c: Double)
+data class Current(val temp_c: Int)
